+++ proxyBankArtifact/src/main/java/fr/adaming/dao/IConseillerDao.java
@@ -62,5 +62,5 @@ public interface IConseillerDao {
 	 * @param prenom
 	 * @return
 	 */
-	public int isExist(String nom,String prenom);
+	public Long isExist(String nom,String prenom);
 }
